Add spacing and remove unwanted padding and margin
Unordered lists have their own margin and padding and these need to be removed to before creating space between list items. A gap was then added to create uniform spacing.

Logo by default takes up entire container in the vertical direction so a manual padding is added to create space between the top and bottom of the logo.